+++ templates/index.html
@@ -5,6 +5,7 @@
     <meta charset="UTF-8">
     <title>SEOライティングツール</title>
     <link rel="stylesheet" href="{{ url_for('static', filename='css/style.css') }}">
+    <script src="//cdnjs.cloudflare.com/ajax/libs/socket.io/4.0.1/socket.io.js"></script>
 </head>
 <body>
